+++ packages/deploy/src/config.ts
@@ -1,6 +1,6 @@
 import { Address } from "./utils";
 import { ProxyAdmin } from "@nocturne-xyz/contracts";
-import { Erc20Config } from "@nocturne-xyz/config";
+import { Erc20Config, ProtocolAddressWithMethods } from "@nocturne-xyz/config";
 import * as JSON from "bigint-json-serialization";
 
 export interface NocturneDeployConfigProperties {
@@ -9,7 +9,8 @@ export interface NocturneDeployConfigProperties {
   subtreeBatchFillers: Address[];
   wethAddress: Address;
   erc20s: [string, Erc20Config][];
-  protocolAllowlist: [string, Address][];
+  protocolAllowlist: [string, ProtocolAddressWithMethods][];
+  leftoverTokenHolder: Address;
   opts?: NocturneDeployOpts;
 }
 
@@ -19,17 +20,8 @@ export interface NocturneDeployConfig {
   subtreeBatchFillers: Address[];
   wethAddress: Address;
   erc20s: Map<string, Erc20Config>;
-  protocolAllowlist: Map<string, Address>;
-  opts?: NocturneDeployOpts;
-}
-
-export interface NocturneDeployConfig {
-  proxyAdminOwner: Address;
-  screeners: Address[];
-  subtreeBatchFillers: Address[];
-  wethAddress: Address;
-  erc20s: Map<string, Erc20Config>;
-  protocolAllowlist: Map<string, Address>;
+  protocolAllowlist: Map<string, ProtocolAddressWithMethods>;
+  leftoverTokenHolder: Address;
   opts?: NocturneDeployOpts;
 }