+++ packages/contracts/CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ### Unreleased
 
+- add functions to ensure Handler balances are zeroed out pre-operation
+- make protocol allowlist per-method, add separate allowlist for tokens
 - Remove ERC721/1155s from all unit and invariant tests
 - Remove ERC721/1155 onReceived hooks in the Teller and Handler
-- Remove the \_receivedAssets mechanism in the Handler now that the onReceived hooks are gone
+- Remove the `_receivedAssets` mechanism in the Handler now that the onReceived hooks are gone
 - Estimate totalNumRefundsToHandle as number of joinsplits + number of refunds assets
 - fix eip712 deposit request typehash ordering
 - update gas comp numbers in `Types.sol` and remove verify gas estimate since not needed